add photo cgrect
