+++ MyMessager_GB/UI/Friends/Persone/View/PersoneCollectionViewCell.swift
@@ -9,11 +9,22 @@ import UIKit
 
 class PersoneCollectionViewCell: UICollectionViewCell {
     static let reuseIdentifier = "PersoneCollectionViewCell"
+    private weak var model: Picture?
+//    var tappedLike: ((_ model: Picture?) -> ())?
+    
     @IBOutlet private weak var picture: UIImageView!
     @IBOutlet private weak var likeControll: LikeControl!
-    func configure(image: UIImage?, countLikes: Int?, isLiking: Bool?){
-        picture.image = image
-        likeControll.countLikes = countLikes ?? 0
-        likeControll.isLiking = isLiking ?? false
+    
+    func configure(picture: Picture){
+        self.model = picture
+        self.picture.image = picture.image
+        likeControll.countLikes = picture.likeCount
+        likeControll.isLiking = picture.isLiking
+        likeControll.addTarget(self, action: #selector(likeTarget), for: .valueChanged)
+    }
+    
+    @objc private func likeTarget(){
+        model?.likeCount = likeControll.countLikes
+        model?.isLiking = likeControll.isLiking
     }
 }
